Admin page snippet done

+++ client/src/components/AdminForm.js
@@ -1,9 +1,10 @@
 import { useEffect, useState } from "react";
 import { useAuthContext } from "../hooks/useAuthContext";
-
+import '../styles/styles.css'
 const AdminForm = () => {
     const [usersList, setUsersList] = useState([]);
     const { user } = useAuthContext();
+    const [errors, setErrors] = useState();
     useEffect(() => {
         const getAllUsers = async () => {
             try {
@@ -26,25 +27,32 @@ const AdminForm = () => {
               }
         }
         getAllUsers();
-    },[])
+    },[usersList])
 
-    const handleDeleteUser = async () => {
+    const handleDeleteUser = async (userid, email, role) => {
         try {
-            const response = await fetch(`http://localhost:3001/api/user/${userId}`, {
+            if(userid === user.userid) {
+                throw new Error('Et voi poistaa itseäsi.');
+            }
+            if(role === 1) {
+                throw new Error('Et voi poistaa toista adminia.');
+            }
+            const response = await fetch(`http://localhost:3001/api/user/${userid}`, {
               method: 'DELETE',
               headers: {
                 'Content-Type': 'application/json',
-                // Lisää tarvittaessa muita otsikoita, kuten esimerkiksi autentikaatioon liittyviä otsikoita
+                'Authorization': `Bearer ${user.token}`
               },
             });
-        
             if (!response.ok) {
               throw new Error(`HTTP error! Status: ${response.status}`);
             }
-        
-            const data = await response.json();
-            console.log('User deleted:', data);
+            if(response.ok){
+                setErrors('Käyttäjä ' + email +  ' poistettiin onnistuneesti.')
+                console.log('User deleted: ', email);
+            }
           } catch (error) {
+            setErrors(error.message)
             console.error('Error deleting user:', error.message);
           }
     }
@@ -53,18 +61,20 @@ const AdminForm = () => {
             <table>
                 <thead>
                     <tr>
-                        <th></th>
+                        <th>Email</th>
                     </tr>
                 </thead>
                 <tbody>
                     {usersList.map((user, index) => (
                     <tr key={index}>
                         <td>{user.email}</td>
-                        <button onClick={() => {}}>Poista käyttäjä: {user.name}</button>
+                        <td><button onClick={() => handleDeleteUser(user.userid, user.email, user.ROLE)}>Poista käyttäjä: {user.name}</button></td>
                     </tr>
                     ))}
                 </tbody>
             </table>
+            <br></br>
+            {errors && <p className="pError">{errors}</p>}
         </div>
       );
 }
